feat: доделал регистрацию, подтверждение почты и логин Добавил страницу профиля автоматизировал сборку проекта, контейнеризировал

+++ src/main/java/ru/itis/raslgab/gowork/repositories/BookingRepo.java
@@ -3,5 +3,5 @@ package ru.itis.raslgab.gowork.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.itis.raslgab.gowork.models.Booking;
 
-public interface BookingRepo extends JpaRepository<Booking, Integer> {
+public interface BookingRepo extends JpaRepository<Booking, Long> {
 }
