+++ src/main/java/ru/itis/raslgab/gowork/repositories/CityRepo.java
@@ -3,5 +3,5 @@ package ru.itis.raslgab.gowork.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.itis.raslgab.gowork.models.City;
 
-public interface CityRepo extends JpaRepository<City, Integer> {
+public interface CityRepo extends JpaRepository<City, Long> {
 }
