+++ src/main/java/ru/itis/raslgab/gowork/repositories/OptionRepo.java
@@ -3,5 +3,5 @@ package ru.itis.raslgab.gowork.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.itis.raslgab.gowork.models.Option;
 
-public interface OptionRepo extends JpaRepository<Option, Integer> {
+public interface OptionRepo extends JpaRepository<Option, Long> {
 }
