+++ src/main/java/ru/itis/raslgab/gowork/repositories/RoomRepo.java
@@ -3,5 +3,5 @@ package ru.itis.raslgab.gowork.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.itis.raslgab.gowork.models.Room;
 
-public interface RoomRepo extends JpaRepository<Room, Integer> {
+public interface RoomRepo extends JpaRepository<Room, Long> {
 }
